Implement Element::scrollIntoView (#463)
* Implement window resizing when virtual keyboard shows
* Fix relese build warnings
* Fix bugs in scrolling
* Move style of size attribute HTMLInputElement into HTMLInputElement::styleForPresentationAttribute
* Implement applying transform in getBoundingClientRect

+++ src/core/dom/MouseEvent.h
@@ -59,7 +59,8 @@ public:
     }
 
     MouseData(unsigned char button, unsigned char buttons, double clientX,
-              double clientY, double screenX, double screenY, int8_t clickCount)
+              double clientY, double screenX, double screenY,
+              int32_t clickCount)
         : m_button(button)
         , m_buttons(buttons)
         , m_clientX(clientX)
